Update analysis tools (#152)
* Extend analysis tool

* Update analysis runner

+++ examples/scripts/diagnostic_tools/doanalysis.py
@@ -69,6 +69,7 @@ def set_o2_analysis(o2_analyses=["o2-analysis-hf-task-d0 --pipeline qa-tracking-
         write_instructions(f"cd {output_path} || exit 1", n=2)
         # Print run dir
         write_instructions(f"pwd", n=2)
+        write_instructions(f"echo Running \"$0\"", n=2)
 
         def get_tagged_output_file(output_file_name):
             return output_file_name.replace(".root", f"_{tag}.root")
@@ -89,11 +90,13 @@ def set_o2_analysis(o2_analyses=["o2-analysis-hf-task-d0 --pipeline qa-tracking-
             if i == o2_analyses[0]:
                 line += f" --aod-file {input_file}"
             if dpl_configuration_file is not None:
-                line += f" --configuration json://{dpl_configuration_file}"
+                line += f" --configuration json://{os.path.normpath(dpl_configuration_file)}"
             if len(o2_analyses) > 1 and i != o2_analyses[-1]:
                 line = f"{line} | \\\n \t"
             else:
                 line = f"{line}"
+            if line.count("configuration") > 1:
+                fatal_msg("Cannot have more than one configuration")
             o2_workflow += line
 
         write_instructions(f"O2Workflow=\"{o2_workflow}\"", n=2)
@@ -106,6 +109,9 @@ def set_o2_analysis(o2_analyses=["o2-analysis-hf-task-d0 --pipeline qa-tracking-
         write_instructions("  eval \"${O2Workflow}\"")
         write_instructions("fi")
 
+        # Print run dir
+        write_instructions("pwd")
+
         for i in ["ERROR", "FATAL", "crash"]:
             write_instructions(
                 f"if grep -q \"\[{i}\]\" {log_file}; then echo \": got some {i}s in '{log_file}'\" && exit 1; fi")
@@ -115,15 +121,43 @@ def set_o2_analysis(o2_analyses=["o2-analysis-hf-task-d0 --pipeline qa-tracking-
             write_instructions(
                 f"[ -f {i} ] && mv {i} {get_tagged_output_file(i)} 2>&1")
 
+        write_instructions(f"date", n=2)
+        write_instructions(f"echo Completed \"$0\"", n=2)
+
         write_instructions("\nexit 0")
     return tmp_script_name
 
 
-def run_o2_analysis(tmp_script_name, remove_tmp_script=False, explore_bad_files=True):
+do_bash_script = False
+bash_parallel_jobs = 1
+
+
+def run_o2_analysis(tmp_script_name,
+                    remove_tmp_script=False,
+                    explore_bad_files=False,
+                    time_it=True):
+    global number_of_runs
     verbose_msg("> starting run with", tmp_script_name)
     cmd = f"bash {tmp_script_name}"
+    if do_bash_script:
+        with open("parallelbash.sh", "a") as fout:
+            with open("parallelbash.sh", "r") as fin:
+                lastline = fin.readlines()[-1]
+                if lastline.startswith("#"):
+                    lastline = int(lastline.strip("#"))
+                else:
+                    lastline = 0
+                fout.write(f"echo Running {lastline}\n")
+                fout.write(f"{cmd} &\n")
+                lastline += 1
+                if lastline % (bash_parallel_jobs+1) == 0:
+                    fout.write(f"wait\n")
+                fout.write(f"\n#{lastline}\n")
+
+        return
+
     if explore_bad_files:
-        if run_cmd(cmd, check_status=True, throw_fatal=False) == False:
+        if run_cmd(cmd, check_status=True, throw_fatal=False, time_it=time_it) == False:
             list_name = os.listdir(os.path.dirname(tmp_script_name))
             for i in list_name:
                 if "ListForRun5Analysis" in i:
@@ -137,7 +171,7 @@ def run_o2_analysis(tmp_script_name, remove_tmp_script=False, explore_bad_files=
             warning_msg("Issue when running",
                         tmp_script_name, "with", list_name)
     else:
-        run_cmd(cmd)
+        run_cmd(cmd, log_file=f"{tmp_script_name}.log", time_it=time_it)
     if remove_tmp_script:
         os.remove(tmp_script_name)
     verbose_msg("< end run with", tmp_script_name)
@@ -165,7 +199,12 @@ def main(mode,
          extra_arguments="",
          resume_previous_analysis=False,
          check_input_file_integrity=True,
-         analysis_timeout=None):
+         analysis_timeout=None,
+         linearize_single_core=True):
+    if do_bash_script:
+        njobs = 1
+        linearize_single_core = True
+
     if len(input_file) == 1:
         input_file = input_file[0]
     else:
@@ -181,10 +220,11 @@ def main(mode,
     if analysis_timeout is not None:
         msg("Using analysis timeout of", analysis_timeout,
             "seconds", color=bcolors.BOKBLUE)
+        analysis_timeout = f"--time-limit {analysis_timeout}"
     else:
-        analysis_timeout = 0
+        analysis_timeout = ""
 
-    o2_arguments = f"-b --shm-segment-size {shm_mem_size} --aod-memory-rate-limit {rate_lim} --readers {readers} --time-limit {analysis_timeout}"
+    o2_arguments = f"-b --shm-segment-size {shm_mem_size} --aod-memory-rate-limit {rate_lim} --readers {readers} {analysis_timeout}"
     o2_arguments += extra_arguments
     if mode not in analyses:
         raise ValueError("Did not find analyses matching mode",
@@ -295,9 +335,27 @@ def main(mode,
                                         resume_previous_analysis=resume_previous_analysis,
                                         write_runner_script=not merge_only))
     if not merge_only:
+        if do_bash_script:
+            with open("parallelbash.sh", "w") as f:
+                f.write(f"#!/bin/bash\n\n")
+                f.write(f"echo \"Start running\"\n\n")
+                f.write(f"date\n\n")
+                f.write("""function trap_ctrlc (){
+                            # perform cleanup here
+                            echo "Ctrl-C caught...performing clean up"
+                            exit 2
+                        }\n\n""")
+                f.write("""trap "trap_ctrlc" 2\n""")
+
         run_in_parallel(processes=njobs, job_runner=run_o2_analysis,
                         job_arguments=run_list, job_message=f"Running analysis, it's {datetime.datetime.now()}",
-                         linearize_single_core=True)
+                        linearize_single_core=linearize_single_core)
+        if do_bash_script:
+            with open("parallelbash.sh", "a") as f:
+                f.write(f"wait\n\n")
+                f.write(f"date\n\n")
+            msg("Now run bash script `bash parallelbash.sh`")
+            return
         if clean_localhost_after_running:
             run_cmd(
                 "find /tmp/ -maxdepth 1 -name localhost* -user $(whoami) | xargs rm -v 2>&1",
@@ -413,8 +471,10 @@ if __name__ == "__main__":
                         action="store_true", help="Flag avoid running the analysis and to merge the output files into one")
     parser.add_argument("--show", "-s",
                         action="store_true", help="Flag to show the workflow of the current tag")
-    parser.add_argument("--no_clean", "-nc",
+    parser.add_argument("--no_clean", "--noclean", "-nc",
                         action="store_true", help="Flag to avoid cleaning the localhost files after running")
+    parser.add_argument("--do_bash_script", "-P",
+                        action="store_true", help="Flag to create a bash script that runs all the tasks in cascade")
     parser.add_argument("--resume_previous_analysis", "--continue_analysis", "--resume_analysis", "--continue",
                         action="store_true",
                         help="Flag to continue the analysis from the input files that have been already built and not overwriting the output results")
@@ -423,6 +483,10 @@ if __name__ == "__main__":
                         help="Flag to avoid checking the input file integrity so as to gain time")
     args = parser.parse_args()
     set_verbose_mode(args)
+
+    # Set bash script mode
+    do_bash_script = args.do_bash_script
+    bash_parallel_jobs = args.njobs
 
     # Load analysis workflows
     workflows = configparser.RawConfigParser()
